# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "GlutenFree documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get all products",
                "description": "Return all products with category, subcategory and images, paginated by 12",
                "operationId": "products",
                "parameters": [
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Product category slug",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "molochnyje-produkty"
                    },
                    {
                        "name": "subcategory",
                        "in": "query",
                        "description": "Product subcategory slug",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "moloko"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": "List of products. Pagination by 12"
                            }
                        }
                    }
                }
            }
        },
        "/products/{slug}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get one product by slug",
                "description": "Return product with category, subcategory and images, paginated by 12",
                "operationId": "products-show",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Product slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "moloko-moloczyj-mir"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": "One product."
                            }
                        }
                    }
                }
            }
        },
        "/user/forum-topics/store": {
            "post": {
                "tags": [
                    "User-Forum-topics"
                ],
                "summary": "Create forum topic",
                "description": "Store posts category",
                "operationId": "store-forum-topic",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreForumTopicRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully created"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreForumTopicRequest": {
                "title": "StoreForumTopicRequest",
                "description": "Class StoreForumTopicRequest",
                "required": [
                    "category_id",
                    "title",
                    "text"
                ],
                "properties": {
                    "category_id": {
                        "title": "category_id",
                        "description": "Forum category id",
                        "type": "integer",
                        "example": "1"
                    },
                    "title": {
                        "title": "title",
                        "description": "Title of user's topic (question)",
                        "type": "string",
                        "maxLength": 255,
                        "minLength": 10,
                        "example": "Ktoś zna gdzie kupić leki?"
                    },
                    "text": {
                        "title": "text",
                        "description": "Text of user's topic (question)",
                        "type": "string",
                        "maxLength": 2000,
                        "minLength": 10,
                        "example": "Szukam pomocy, nie mogę znaleźć lek xxx, czy ktoś kiedyś go kupował?"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer \"token\")",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        []
    ]
}